feat: show a skill's connected apps across the skill surfaces
Skills that declare Composio toolkit slugs in their frontmatter now show
those apps everywhere the skill appears: logo chips on picker/chat skill
cards, pips on the installed-skills strip rows, named badges in the
skill edit modal (new integrationsSlot prop in @houston-ai/skills), and
a logo row on the chat invocation card.

AgentIntegrationChips becomes the shared IntegrationChips (reusing
AppLogo, real-name tooltips); IntegrationBadges is extracted from the
store detail dialog and shared with the edit modal; AppLogo gains an xs
size. Slugs are normalized once via skillIntegrationSlugs (trim,
lowercase, dedupe) at every callsite. New i18n key
skills:detail.integrations in en/es/pt.

Fixes HOU-794
